# expression: implement vectorized evaluation for `builtinArithmeticDivideRealSig`